Realmode portion of the code no longer uses GCC-IA16 and LD-IA16 to build. It now uses regular GCC with the -m16 flag. This means that now, the realmode code is actually 32-bit code with operand-length prefixes. The reason for this change is that GCC-IA16 produces pure 16-bit code, which means we can't use 32-bit registers, which are necessary for some BIOS functions.

+++ arch/x86/realmode/load_kernel.c
@@ -52,7 +52,7 @@ void load_section()
 
     // copy the section from its temporary location to its permanent location
     // using the 32 bit protected mode "copy_section" assembly routine
-    switch_protected_mode((uint16)copy_section);
+    switch_protected_mode((uint32)copy_section);
 }
 
 /**
